fix menu spring problem

+++ ProgrammingTasks/src/main/java/by/bsu/fpmi/kolyadkodarya/dao/CategoryDao.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public interface CategoryDao
 {
-    Category getById(Long id);
+    Category getById(int id);
     void update(Category category);
     Category save(Category category);
     List<Category> listCategories();
