+++ ProgrammingTasks/src/main/java/by/bsu/fpmi/kolyadkodarya/services/CategoryService.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public interface CategoryService
 {
-    Category getById(Long id);
+    Category getById(int id);
 
     void update(Category category);
